src/HardwareSensorSystem.Security/Controllers/UserController.cs
using HardwareSensorSystem.Security.Models;
using HardwareSensorSystem.Security.ViewModels;
using Microsoft.AspNetCore.Identity;
using Microsoft.AspNetCore.Mvc;
using System.Linq;
using System.Threading.Tasks;

namespace HardwareSensorSystem.Security.Controllers
{
    public class UserController : Controller
    {
        private UserManager<ApplicationUser> _userManager;

        public UserController(UserManager<ApplicationUser> userManager)
        {
            _userManager = userManager;
        }

        public Task<IActionResult> GetAll()
        {
            var users = _userManager.Users.Select(u => u.ToViewModel());
            return Task.FromResult<IActionResult>(Ok(users));
        }
    }
}
